Make Span Event a top level class
