+++ misc/tracing-extensions/modules/ballerina-choreo-extension/src/main/java/org/ballerinalang/observe/trace/extension/choreo/client/ChoreoClient.java
@@ -33,6 +33,7 @@ import org.ballerinalang.observe.trace.extension.choreo.logging.LogFactory;
 import org.ballerinalang.observe.trace.extension.choreo.logging.Logger;
 import org.ballerinalang.observe.trace.extension.choreo.model.ChoreoMetric;
 import org.ballerinalang.observe.trace.extension.choreo.model.ChoreoTraceSpan;
+import org.ballerinalang.observe.trace.extension.choreo.model.SpanEvent;
 
 import java.util.List;
 import java.util.Objects;
@@ -201,7 +202,7 @@ public class ChoreoClient implements AutoCloseable {
                 }
 
                 if (traceSpan.getEvents() != null) {
-                    for (ChoreoTraceSpan.SpanEvent spanEvent: traceSpan.getEvents()) {
+                    for (SpanEvent spanEvent: traceSpan.getEvents()) {
                         traceSpanBuilder.addCheckpoints(TelemetryOuterClass.Checkpoint.newBuilder()
                                 .setTimestamp(spanEvent.getTime())
                                 .setModuleID(spanEvent.getModuleID())
